continue refining electrostatic base class
