+++ Source/FieldSolver/ElectrostaticSolvers/ElectrostaticSolver.cpp
@@ -18,10 +18,6 @@ ElectrostaticSolver::ElectrostaticSolver (int nlevs_max)
     // {
 
     // }
-
-    // Create an instance of the boundary handler to properly set boundary
-    // conditions
-    m_poisson_boundary_handler = std::make_unique<PoissonBoundaryHandler>();
 }
 
 ElectrostaticSolver::ComputeSpaceChargeField (
